requirements.txt
python-dotenv>=1.0.1
requests>=2.31.0
pandas>=1.5.0
openai>=1.3.5
sqlparse
streamlit-code-editor
duckdb>=1.3.2